backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
pytest>=8.0.0
httpx>=0.27.0
openai>=1.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
PyJWT>=2.8.0
